refactor: renaming and small fixes
