+++ web-registry/src/components/organisms/CreditBatchesSection/CreditBatchesSection.tsx
@@ -4,10 +4,10 @@ import Section from 'web-components/lib/components/section';
 import { Title } from 'web-components/lib/components/typography';
 
 import { CreditBatches } from '../CreditBatches/CreditBatches';
-import { useProjectTopSectionStyles } from './ProjectMiddleSection.styles';
-import { ProjectMiddleSectionProps } from './ProjectMiddleSection.types';
+import { useProjectTopSectionStyles } from './CreditBatchesSection.styles';
+import { ProjectMiddleSectionProps } from './CreditBatchesSection.types';
 
-function ProjectMiddleSection({
+function CreditBatchesSection({
   offChainProject,
   batchData,
   paginationParams,
@@ -18,7 +18,7 @@ function ProjectMiddleSection({
   return (
     <Section classes={{ root: classes.section }}>
       {batchData?.batches && batchData.batches.length > 0 && (
-        // spacing here based on paddding-top for `<Section />` component
+        // spacing here based on padding-top for `<Section />` component
         <Box sx={{ mt: { xs: 17.75, sm: 22.25 } }}>
           <Title variant="h3" sx={{ pb: 8 }}>
             Credit Batches
@@ -39,4 +39,4 @@ function ProjectMiddleSection({
   );
 }
 
-export { ProjectMiddleSection };
+export { CreditBatchesSection };
